requirements.txt
playwright==1.55.0
azure-functions
azure-storage-blob
